Moved block selector dialog code to its own folder
